# Slightly improve Latitude.validate_angles performance